requirements.txt
flask 
joblib 
pandas 
pytesseract 
pillow 
thefuzz
